Rename foo => main, and flesh out main#read.
Add a migration which makes a database table called "comment". `Comment` is a
reserved keyword (already used by rails itself), so the "comment" table is
available through the class `Comment_`. (see app/models/comment_.rb)

Rewrite routes file. It's not in a finished form. The main#create route is also
not in a finished form.