Add a simple proc file.
This proc file can be created and destroyed by inserting and removing a module
from the kernel. The proc file does nothing useful (it prints out a series of
even numbers), but it can serve as an excellent template for studying how
virtual files work and for creating new proc files.