Rewrite and expand vsftpd configs.
Change the naming scheme for vsftpd configs to be more concise. Trim down the
huge number of comments in the configs (no need to replicate the man page), and
make the remaining configs a bit more useful.

The new config file allows for SSL-encrypted, read-write connections to be made
by explicitly named local users.