Update libvirt setup and teardown scripts
dnsmasq makes note of the hostnames of DHCP clients and stores that
information in its leases file. This allows a NATted VM to perform DNS
lookups for the names of other VMs. For example, if a VM performs a DNS
lookup for "nslookup fedora-24-pulp-2-10", dnsmasq will consult its DHCP
lease file and either return an IP address or an error as appropriate.

Unfortunately, this doesn't help the VM host perform DNS lookups for its
NATted VMs. When the host performs a DNS lookup, it sends a request to
one of the DNS servers listed in `/etc/resolv.conf`, none of which are
aware of the DHCP leases file maintained by dnsmasq.

Solve this issue by ensuring that dnsmasq is listed in the
`/etc/resolv.conf` file. The solution makes use of resolvconf(8), and
should therefore be well-behaved.