Update btrfs-scrub.service
systemd services that have a type of "simple" or "oneshot" are expected
to stay in the foreground. Update the `btrfs-scrub.service` unit file so
that the `btrfs scrub start ...` command stays in the foreground.

Also, change the service type from "oneshot" to "simple." The former
blocks subsequent services from starting. Consequently, it may be
subject to being killed after 90 seconds.