Ensure variable is handled correctly
Performing an "if" check on the literal value of a variable is a bad
idea: if that variable contains special characters, then it can break
the check, and if the variable intentionally has a false value, then it
breaks the user's intentions.

Drop an "if" check on the literal value of a variable. Let the check be
against whether or not the variable is defined.