Rename role: beech-nftables -> nftables
The "nftables" role is generic enough to be applied to most end user
systems. It's not suitable for special cases like routers and libvirt
servers. Rename the beech-nftables role to nftables to reflect this
fact, and adjust host definitions to ensure that the role is applied to
the correct systems.