Drop Ansible host vars for host, add .gitignore
Ansible currently has a host_vars file for beech.ichimonji10.name with
the following contents:

    ansible_connection: local

This is valid and correct when the host running the Ansible playbook
really is beech.ichimonji10.name, but incorrect when some other host is
running this Ansible playbook. (e.g. What if butternut ran the playbook?
It would think itself to be beech.)

Drop the host_vars file for this host. Add a `.gitignore` file matching
*all* files in the `host_vars` directory. Why keep this empty directory
around? Because it seems likely that transient host vars (like
`ansible_connection`) will need to be specified, and the `host_vars`
directory provides a good mechanism for doing so; and because any host
vars that are specified should not be added to version control by
accident. (Files matched by a `.gitignore` file can still be added to
version control with `git add --force`.)