Ansible: Update `netctl-*` roles
When a netctl profile is modified, the corresponding systemd unit file
should be re-made. The appropriate command for this is `netctl reenable
...`, not `netctl enable`. The latter will fail if a corresponding unit
file already exists. Use `netctl reenable ...`.

When a netctl profile is modified, systemd should reload the
corresponding unit file so that the changes are taken into effect.
Update the `netctl-*` roles so that this happens.