Re-organize inventory groups
Let the hosts in the inventory file have the following groups:

* workstations
* servers
* routers
* vm-hosts
* btrfs-hosts

The "servers" group is new, and "vm-hosts" was formerly known as
"vmhosts."

Place descriptions of the groups into the inventory file.

Use Ansible's "patterns" feature to apply plays to interesting sets of
hosts. (An example of a filter is `servers:!routers`.)