Perform fewer user changes in borg-backup role
Currently, the borg-backup role is designed to be invoked with:

    become_user: some_user

This means that there are several declarations defining which user
should currently be executing a task:

1. The `become_user` defined by `ansible.cfg` or similar.
2. The `become_user` declaration in the ansible playbook that calls the
   borg-backup role.
3. The `become_user` declaration attached to tasks in the `borg-backup`
   role.

This is a bit much. Simplify things, so that the following declarations
define which user might execute a task:

1. The `become_user` defined by `ansible.cfg` or similar.
2. The `become_user` declaration attached to tasks in the `borg-backup`
   role.

This should make it easier to work with the borg-backup role in the
future.