JJB: Update script to for creating venv
Let the script install both pip and jenkins-job-builder without the
`--upgrade` flag. Formerly, the script installed pip in upgrade mode and
jjb normally, which is a weird pairing.

The script could also have been configured to install both tools with
the `--upgrade` flag, but that solution would have produced a script
that is slower and non-idempotent.