Ansible: Add role for creating users
All user creation is currently done with inline tasks in the
`manage-users.yml` playbook. Move user creation into a dedicated
`create-users` role. This role is special: it's the first one in which
tests for group membership are placed in roles. Formerly, all tests for
group membership occurred in the playbooks that call roles.