Ansible: Overhaul user management
Drop the `create-users` and `customize-user-ichimonji10` roles. Instead,
create one role per user. Currently, they are:

* `user-googoo`
* `user-ichimonji10`

These roles may conditionally depend on other roles. For example, this
is in `meta/main.yml` for `user-ichimonji10`:

    dependencies:
      # Creates ichi-machines group.
      - {role: msmtp}
      # Creates kvm and libvirt groups.
      - {role: libvirtd, when: "'vm-hosts' in group_names"}

At first glance, this seems like a regression. Shouldn't all logic
related to which roles are applied to which hosts be in playbooks? In
practice, it's quite nice. User roles like the late
`customize-user-ichimonji10` already had knowledge of groups, and this
organization ends up being more natural. Just apply a `user-*` role to
whichever hosts are appropriate, and that role knows how to deal with
the different types of hosts it is applied to.

For more thoughts on this topic, see edunham's blog post on [conditional
role
dependencies](http://edunham.net/2015/09/10/ansible_conditional_role_dependencies.html).
Beware, though, that that blog post considers variables provided by
Ansible itself, whereas this commit considers variables defined in the
inventory, so the use cases are similar but different.