Ansible: Make calls to pacman role more concise
Don't use for loops to install packages one-by-one. Instead, pass all
packages that pacman should install as a list to the pacman module. It's
unclear whether or not the pacman module internally handles these lists
of packages efficiently, or whether it just performs a for loop
internally. Hopefully so. But even if not, passing a list to the pacman
module means that future revisions to the module will be able to act
more efficiently.