Ansible: Let desktop-env. role use AUR helper
Update the "desktop-environment" role so that packages available through
the AUR are installed with an AUR helper instead of the pacman module.
(An AUR package can only be installed with the pacman module if that
package is installed. It's nearly pointless, unless one wants to know in
an un-graceful way when an AUR package isn't installed.)