Ansible: Correctly install AUR packages
Many roles include calls to the pacman module that look like this:

    - name: Install foo
      pacman:
        name: foo

This works when foo is available from the official repositories, but it
breaks when foo isn't. Many roles include code like this, where foo is
only available from the AUR. This will break, as pacman doesn't install
packages from the AUR.

Update code so that these roles depend on the `user-pacaur-builder`
role, and so that these roles install packages using pacaur.

Drop the reference to `user-pacaur-builder` from `site.yml`. This role
should be pulled in via dependencies as needed.