Update borg-backup role
Add a readme.

Let the role's tasks file assert that required variables are present. If
a user calls a role and fails to pass in required information, they
should be made aware of their mistake in a rude fashion.

In the calls to this role, tighten variable scope.

Drop dependencies on user roles like `user-ichimonji10`. It makes no
sense for a role to have a parameter called `borg_backup_user`, and then
to include explicit instructions to create all the users that might be
passed in via that variable.