Don't add ichimonji10 to "kvm" group
Among other things, the clone-customize-domains script does two things:

1. Clone VMs with the virt-clone executable.
2. Customize the new disk images with the virt-sysprep executable.

In order to successfully complete the second step, the user running
clone-customize-domains must have permissions to access disk images.
This is currently done by setting the group of each disk image file to
"kvm." So long as the user belongs to the "kvm" group, this works.

There's a downside to this appraoch: unnecessary privilege escalation.
Why should a user belong to the "kvm" group just so that they can access
a few disk images? Another approach is for the script to set the group
of each disk image file to "$(id --group)".

Implement the latter approach. Don't add ichimonji10 to the "kvm" group.
Ensure that users of this script can access disk image files by setting
the group on each file to "$(id --group)".