Let pine listen for ssh connections on port 22
When a packet is sent to amelanchier on port 30022, it's forwarded to
pine on that same port. This is annoying, because pine must have a
non-standard sshd configuration for this to work.

Fix this situation. Let traffic to amelanchier on port 30022 be
forwarded to pine on port 22.

Note that the following resources did *not* help:

* https://wiki.nftables.org/wiki-nftables/index.php/Maps
* https://wiki.nftables.org/wiki-nftables/index.php/Multiple_NATs_using_nftables_maps

nftables threw errors whenever I used the strategies outlined therein. I
instead used the strategies outlined in nft(8), specifically the "NAT
STATEMENTS" section.