Rename several templates in the borg-backup role
Drop the `.j2` suffix from several templates. This allows syntax
highlighting to work correctly in editors. And the fact that the file is
in a "templates" directory makes it obvious that the files are
templates.

Add a `.sh` suffix to a shell script. No need to be obscure about the
type of a file.