Improve dynamic-dns role
Add a readme.

Let the role behave correctly even when `dynamic_dns_urls` is not
defined. When this is the case, the task which installs
`/usr/local/bin/dynamic-dns` will skip. And if this file doesn't exist,
then the unit which starts and enables the units that call said script
will skip.