Move btrfs_subvolumes from playbook to host_vars
Playbooks are far easier to understand when they're only concerned with
mapping hosts to roles, and when they're not concerned with defining
variables. In addition, letting variable definitions be concentrated in
the `host_vars` and `group_vars` directories lets variable definitions
be easier to find and read through.

Move the `btrfs_subvolumes` variable, used by the `btrfs-subvolume`
role, from the `site.yml` playbook to file(s) in the `host_*`
directories.a