Move syncthing_usernames out of playbook
Playbooks are easier to understand when they're only concerned with
mapping hosts to roles, and when they're not concerned with defining
variables. In addition, letting variable definitions be concentrated in
the `host_vars` and `group_vars` directories lets variable definitions
be easier to find and read through.

Move the `syncthing_usernames` variable, used by the `syncthing` role,
from the `site.yml` playbook to file(s) in the `host_*` directories.