Change how crda role works
The wireless regulatory domain can be set in the
`/etc/conf.d/wireless-regdom` configuration file. The crda role
currently edits this file using the lineinfile module. This works, but
is not ideal: the use of this module implies that two different
applications are editing the same file. Management is simplified by
having just one application control this file. Drop the lineinfile
module in favor of the copy module.

This change fixes one ansible-review warning.