Use script module instead of shell module
One can invoke a shell script on a remote system with the shell module:

```yaml
- name: Run a script
  shell: |
    whoami
    pwd
```

This works, but it doesn't scale well. If one wants to execute a complex
script, it would be much nicer if this script could be placed into its
own dedicated file. This would allow easier syntax checking, linting,
stand-alone testing, and so on. The script module lets one do this:

```yaml
- name: Run a script
  script: my-script.sh
```

Do this in the aur role.