Be less aggressive about exporting $TERM
tmux doesn't properly report its ability to display colors, so
`~/.bashrc` includes a kludge that forces 256-color mode when tmux is
running: the TERM environment variable is set. It would be even better
to configure tmux to properly advertise its capabilities, but this
solution is OK enough for general use.

However, this doesn't mean that TERM should be forced into 256-color
mode in other circumstances. For some reason, the bashrc file is set to
do this. Drop this apparently unnecessary TERM-setting logic.