Let the aur user be nicer to mirrors
The "aur" role automatically updates the packages it provides on a daily
basis, at exactly midnight. While this is fine on a small scale, it's
bad on a large scale: a large number of clients all hitting mirrors at
exactly the same time is a great way to swamp them. Add a
`RandomizedDelaySec` to the .timer unit that updates AUR packages, so as
to be nicer to these mirrors.