Change when reflector-update.timer triggers
`reflector-update.timer` is configured to trigger only on Monday,
Wednesday and Friday. This is too infrequent. Failing to update the list
of mirrors for three full days can lead to the use of slow or outdated
mirrors. Re-configure the timer to run daily. This is more useful, and
is hopefully still not an undue burden on the mirrors.

`reflector-update.timer` is configured to trigger exactly at midnight.
At a large scale, this has the potential to be punishing to mirrors. Add
a several minute long randomized delay.