Rewrite the aur role's unit definitions
Drop the per-user units from the aur user's home directory. Instead,
install the aur package update units in `/etc/systemd/system/`, and
ensure that the units execute as the aur user. This allows for easier
system management and fewer lingering systemd instances.

Don't let the aur user belong to the ichi-machines group. This means
that it can no longer send emails. Drop the aur role's dependency on the
msmtp role, which creates the ichi-machines group. All of this is
unnecessary, as system-wide monitoring is enough.

Give the aur user a shell of /usr/bin/nologin. There's no reason for
anything more.

Let the aur user be a system user.