Drop borg-related groups from inventory
When borg backs up files, it's because something else is generating
those files. For example, perhaps a host is a workstation, and that
workstation's users generate personal files. In this case, it makes
sense to call the `borg-backup-users` role for all workstations, and
having an additional group that includes all the same hosts as the
workstations group is redundant.